requirements.txt
bs4
requests
pandas
googlesearch-python
google
gnews